fix type
