--- src/cli.ts
@@ -1,3 +1,0 @@
-import { PortalNode } from ".";
-
-new PortalNode({ port: 60000 });
