--- src/examples/express/src/b.ts
@@ -1,3 +1,0 @@
-import portalnode from "./portalnode";
-
-portalnode(60001, "http://localhost:60000");
