--- src/examples/express/src/index.ts
@@ -1,3 +1,0 @@
-import portalnode from "./portalnode";
-
-portalnode(60000);
