+++ src/examples/socket.io/guest/index.ts
@@ -1,15 +1,10 @@
 import sha1 from "sha1";
 import client from "socket.io-client";
-import Kademlia from "../../kademlia";
-import { PeerModule } from "../../kademlia/modules/peer/webrtc";
-import Peer from "../../kademlia/modules/peer/base";
 import Event from "rx.mini";
-import { Option } from "../../kademlia/ktable";
-import { KvsModule } from "../../kademlia/modules/kvs/base";
+import { Kademlia, PeerModule, KvsModule, Peer } from "../../..";
 
 type Options = {
   target: { url: string; port: number };
-  kadOption?: Partial<Option>;
 };
 
 const Request = (clientKid: string) => {
@@ -36,11 +31,7 @@ type actions = Offer | Request | Answer;
 
 export default class GuestNode {
   kid = sha1(Math.random().toString()).toString();
-  kademlia = new Kademlia(
-    this.kid,
-    { peerCreate: PeerModule, kvs: KvsModule() },
-    this.opt.kadOption
-  );
+  kademlia = new Kademlia(this.kid, { peerCreate: PeerModule, kvs: KvsModule });
   peers: { [key: string]: Peer } = {};
   onConnect = new Event();
